+++ src/demos/findex/users.ts
@@ -1,0 +1,74 @@
+import { v4 as uuidv4 } from "uuid";
+import { USERS } from "./usersDataset";
+
+
+export interface User {
+  id: string, firstName: string, lastName: string, phone: string, email: string, country: string, region: string, employeeNumber: string, security: string, enc_uid: string
+}
+
+export class Users {
+  private _users: User[];
+
+  constructor() {
+    this._users = this.generateUsers();
+  }
+
+  generateUsers(): User[] {
+    const users: User[] = [];
+    USERS.map((val: any) => {
+      const user: User = {
+        id: uuidv4(),
+        firstName: val.firstName,
+        lastName: val.lastName,
+        region: val.region,
+        country: val.country,
+        employeeNumber: val.employeeNumber,
+        email: val.email,
+        phone: val.phone,
+        security: val.security,
+        enc_uid: "",
+      }
+      // Add User objet here
+      users.push(user)
+    })
+    return users;
+  }
+
+  getUsers(): User[] {
+    return this._users;
+  }
+
+  getUsersById(uids: string[]): User[] {
+    return this._users.filter(element => {
+      return (uids.indexOf(element.id) > -1)
+    });
+  }
+
+  getFirstUsers(): { firstName: string, lastName: string, phone: string, email: string, country: string, region: string, employeeNumber: string, security: string }[] {
+    const firstUsers: { firstName: string, lastName: string, phone: string, email: string, country: string, region: string, employeeNumber: string, security: string }[] = [];
+    for (let i = 0; i < 5; i++) {
+      firstUsers.push({
+        firstName: this._users[i].firstName,
+        lastName: this._users[i].lastName,
+        phone: this._users[i].phone,
+        email: this._users[i].email,
+        country: this._users[i].country,
+        region: this._users[i].region,
+        employeeNumber: this._users[i].employeeNumber,
+        security: this._users[i].security,
+      });
+    }
+
+    return firstUsers;
+  }
+
+  upsertUserEncUidById(id: string, encryptedUid: { enc_uid: string }): void {
+    for (let i = 0; i < this._users.length; i++) {
+      if (this._users[i].id !== id) {
+        continue;
+      }
+      this._users[i].enc_uid = encryptedUid.enc_uid;
+      break;
+    }
+  }
+}