+++ src/interface/findex/master_keys.ts
@@ -1,0 +1,3 @@
+export interface MasterKeys {
+  k: string, k_star: string
+}