+++ tests/interface/findex/redis/findex.test.ts
@@ -1,0 +1,38 @@
+import { PolicyAxis } from "../../../../src/crypto/abe/keygen/policy";
+import { masterKeysFindex } from "../../../../src/demos/findex/keys";
+import { RedisDB } from "../../../../src/demos/findex/redis/db";
+import { Users } from "../../../../src/demos/findex/users";
+import { Findex } from "../../../../src/interface/findex/findex";
+
+
+test('upsert+search', async () => {
+    const redisDb = new RedisDB("localhost", 6379);
+    //
+    // Upsert Indexes
+    //
+    try {
+        redisDb.initInstance();
+        const findexDemo = new Findex(redisDb);
+
+        const users = new Users();
+        expect(users.getUsers().length).toBe(99)
+
+        await redisDb.instance.flushAll();
+        await findexDemo.upsertUsersIndexes(masterKeysFindex, users, "id");
+
+        const entries = await redisDb.getEntryTableEntries();
+        const chains = await redisDb.getChainTableEntries();
+        expect(entries.length).toBe(577)
+        expect(chains.length).toBe(792)
+
+        //
+        // Search words
+        //
+        const queryResults = await findexDemo.searchWithLogicalSwitch(masterKeysFindex, "france spain", false, 1000);
+        expect(queryResults.length).toBe(60);
+        redisDb.instance.quit()
+    } catch (error) {
+        redisDb.instance.quit()
+        throw new Error("Redis test failed: " + error);
+    }
+})